Fix implicit signed int casting

+++ include/kr8md/operators/construct.h
@@ -7,7 +7,7 @@ namespace kr8md
 {
 #ifdef __SSE4_2__
     template <>
-    KR8MD_INLINE puint32_4::Pak(const uint32_t &value) : intrinsic(_mm_set1_epi32(value)){};
+    KR8MD_INLINE puint32_4::Pak(const uint32_t &value) : intrinsic(_mm_set1_epi32(static_cast<int32_t>(value))){};
 
     template <>
     KR8MD_INLINE pint32_4::Pak(const int32_t &value) : intrinsic(_mm_set1_epi32(value)){};
@@ -18,7 +18,7 @@ namespace kr8md
 
 #ifdef __AVX2__
     template <>
-    KR8MD_INLINE puint32_8::Pak(const uint32_t &value) : intrinsic(_mm256_set1_epi32(value)){};
+    KR8MD_INLINE puint32_8::Pak(const uint32_t &value) : intrinsic(_mm256_set1_epi32(static_cast<int32_t>(value))){};
 
     template <>
     KR8MD_INLINE pint32_8::Pak(const int32_t &value) : intrinsic(_mm256_set1_epi32(value)){};
